@@ -1,6 +1,6 @@
 {
   "name": "gloc",
-  "version": "9.0.6",
+  "version": "9.0.7",
   "engines": {
     "node": "16.19.0",
     "npm": "0.39.3"
@@ -31,6 +31,7 @@
     "build": "vite build",
     "preview": "vite preview",
     "=== EXTENSION ===": "",
+    "i": "npm i --legacy-peer-deps -ddd",
     "pack:prod": "webpack --config webpack.config.prod",
     "pack:dev": "webpack --config webpack.config.dev",
     "zip": "grunt",
@@ -46,11 +47,15 @@
     }
   },
   "dependencies": {
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/styles": "^4.11.4",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@mui/icons-material": "^5.15.1",
+    "@mui/material": "^5.14.18",
+    "@mui/styled-engine": "^5.15.2",
+    "@mui/styles": "^5.15.1",
     "@types/node": "^22.10.3",
-    "@types/react": "^16.9.13",
-    "@types/react-dom": "^16.9.4",
+    "@types/react": "npm:types-react@^19.0.0-rc.1",
+    "@types/react-dom": "npm:types-react-dom@^19.0.0-rc.1",
     "@types/styled-components": "^4.4.0",
     "@typescript-eslint/parser": "^5.62.0",
     "@vitejs/plugin-react": "^4.3.4",
@@ -62,9 +67,9 @@
     "prettier": "^2.8.8",
     "prettier-eslint": "^16.1.1",
     "prettier-eslint-cli": "^8.0.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "styled-components": "^4.4.1",
+    "react": "^19.0.0-rc.1",
+    "react-dom": "^19.0.0-rc.1",
+    "styled-components": "^5.3.0",
     "vite-plugin-checker": "^0.8.0"
   },
   "devDependencies": {
