@@ -1,7 +1,5 @@
-/* import { translateElements } from './utils/translateElements';
-import { MESSAGE_IDS } from './consts/index'; */
 import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import Popup from './modules/Popup/index';
 
@@ -11,7 +9,18 @@ class PopupPage extends React.PureComponent {
   }
 }
 
-const placeToMount =
-  document.getElementById('root'); /*  || document.createElement('div') */
+const container = document.getElementById('root');
 
-ReactDOM.render(<PopupPage />, placeToMount);
+if (container) {
+  const root = createRoot(container);
+
+  root.render(<PopupPage />);
+} else {
+  const newContainer = document.createElement('div');
+  newContainer.id = 'root';
+  document.body.appendChild(newContainer);
+
+  const root = createRoot(newContainer);
+
+  root.render(<PopupPage />);
+}
