@@ -1,14 +1,12 @@
 import * as React from 'react';
 
-import Grid from '@material-ui/core/Grid';
-import Typography from '@material-ui/core/Typography';
-
 import appConfig from '../../../Common/helpers/appConfig';
+import { Grid, Typography } from '@mui/material';
 
 export default class PopupPage extends React.PureComponent {
   render() {
     return (
-      <Grid container direction="row" justify="space-between">
+      <Grid container direction="row" style={{ justifyItems: 'space-between' }}>
         <Grid item>
           <img
             alt="Extension logo"
